simplified SPs

+++ Core/register.h
@@ -40,7 +40,7 @@ namespace	mBrane{
 		template<class	Register>	class	Array{
 		protected:
 			Register	*_array;
-			uint32		_count;
+			uint16		_count;
 		public:
 			Array();
 			~Array();
